adding label prop and removing label  element and renaming file
